requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
numpy
pypdf